pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-trading-bot"
version = "0.1.0"
description = "Hybrid crypto trading bot with market, sentiment, regime, and risk layers."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "jinja2>=3.1.4",
  "pydantic>=2.8.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
live = [
  "ccxt>=4.4.0",
  "praw>=7.8.0",
  "mysql-connector-python>=9.0.0",
  "google-generativeai>=0.8.0",
  "yfinance>=0.2.50",
]

[project.scripts]
cryptobot-cli = "cryptobot.main:main"
cryptobot-web = "cryptobot.webapp:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
